Dev troye

+++ TOMB5/game/joby.h
@@ -4,7 +4,7 @@
 void inject_joby(bool replace);
 
 void KlaxonTremor();
-static int CheckCableBox(PHD_VECTOR* pos, short size);
+static long CheckCableBox(PHD_VECTOR* pos, short size);
 void ControlElectricalCables(short item_number);
 void WreckingBallCollision(short item_num, ITEM_INFO* laraitem, COLL_INFO* coll);
 void ControlWreckingBall(short item_number);
